requirements.txt
asgiref==3.7.2
Django==4.2.7
graphql-core==3.2.3
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
strawberry-graphql==0.213.0
strawberry-graphql-django==0.24.1
typing_extensions==4.8.0
